Display terminated placements
