+++ frontend/src/lib-customizations/espoo/citizen/assets/i18n/sv.tsx
@@ -1739,6 +1739,7 @@ const sv: Translations = {
         'Irtisanoessasi paikkaa huomaathan, että mahdollinen siirtohakemus poistuu viimeisen läsnäolopäivän jälkeen. Jos tarvitset lapsellesi myöhemmin paikan, sinun tulee hakea sitä uudella hakemuksella.',
       until: (date: string) => `voimassa ${date}`,
       choosePlacement: 'Valitse paikka, jonka haluat irtisanoa',
+      terminatedPlacements: 'Olet irtisanonut paikan',
       lastDayInfo:
         'Viimeinen päivä, jolloin lapsesi tarvitsee paikkaa. Paikka irtisanotaan päättymään tähän päivään.',
       lastDayOfPresence: 'Viimeinen läsnäolopäivä',
